chore: 游戏逻辑改造，fp -> oop
